Disable the "Ok" and "Cancel" button after they're pushed
At the moment after the input values are submitted once the SW doesn't display updated values when the inputs are changed.
Disabling the "Ok" button after it's pressed forces the user to restart the program if they want to calculate new values.

This will be reverted once the issue is fixed.